Refactor test package
