+++ component/sqs/integration_test.go
@@ -6,14 +6,16 @@ package sqs
 import (
 	"context"
 	"encoding/json"
+	"fmt"
 	"testing"
 	"time"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
+	awsConfig "github.com/aws/aws-sdk-go-v2/config"
+	"github.com/aws/aws-sdk-go-v2/credentials"
 	"github.com/aws/aws-sdk-go-v2/service/sqs"
 	patronsqscli "github.com/beatlabs/patron/client/sqs"
 	"github.com/beatlabs/patron/correlation"
-	testaws "github.com/beatlabs/patron/test/aws"
 	"github.com/opentracing/opentracing-go/ext"
 	"github.com/prometheus/client_golang/prometheus/testutil"
 	"github.com/stretchr/testify/assert"
@@ -35,9 +37,9 @@ func Test_SQS_Consume(t *testing.T) {
 	const queueName = "test-sqs-consume"
 	const correlationID = "123"
 
-	api, err := testaws.CreateSQSAPI(region, endpoint)
+	api, err := createSQSAPI(region, endpoint)
 	require.NoError(t, err)
-	queue, err := testaws.CreateSQSQueue(api, queueName)
+	queue, err := createSQSQueue(api, queueName)
 	require.NoError(t, err)
 
 	sent := sendMessage(t, api, correlationID, queue, "1", "2", "3")
@@ -122,4 +124,55 @@ func sendMessage(t *testing.T, api *sqs.Client, correlationID, queue string, ids
 	}
 
 	return sentMessages
+}
+
+func createSQSQueue(api SQSAPI, queueName string) (string, error) {
+	out, err := api.CreateQueue(context.Background(), &sqs.CreateQueueInput{
+		QueueName: aws.String(queueName),
+	})
+	if err != nil {
+		return "", err
+	}
+
+	return *out.QueueUrl, nil
+}
+
+type SQSAPI interface {
+	CreateQueue(ctx context.Context, params *sqs.CreateQueueInput, optFns ...func(*sqs.Options)) (*sqs.CreateQueueOutput, error)
+	ReceiveMessage(ctx context.Context, params *sqs.ReceiveMessageInput, optFns ...func(*sqs.Options)) (*sqs.ReceiveMessageOutput, error)
+}
+
+func createSQSAPI(region, endpoint string) (*sqs.Client, error) {
+	cfg, err := createConfig(sqs.ServiceID, region, endpoint)
+	if err != nil {
+		return nil, err
+	}
+
+	api := sqs.NewFromConfig(cfg)
+
+	return api, nil
+}
+
+func createConfig(awsServiceID, awsRegion, awsEndpoint string) (aws.Config, error) {
+	customResolver := aws.EndpointResolverWithOptionsFunc(func(service, region string, options ...interface{}) (aws.Endpoint, error) {
+		if service == awsServiceID && region == awsRegion {
+			return aws.Endpoint{
+				URL:           awsEndpoint,
+				SigningRegion: awsRegion,
+			}, nil
+		}
+		// returning EndpointNotFoundError will allow the service to fallback to it's default resolution
+		return aws.Endpoint{}, &aws.EndpointNotFoundError{}
+	})
+
+	cfg, err := awsConfig.LoadDefaultConfig(context.TODO(),
+		awsConfig.WithRegion(awsRegion),
+		awsConfig.WithEndpointResolverWithOptions(customResolver),
+		awsConfig.WithCredentialsProvider(aws.NewCredentialsCache(credentials.NewStaticCredentialsProvider("test", "test", ""))),
+	)
+	if err != nil {
+		return aws.Config{}, fmt.Errorf("failed to create AWS config: %w", err)
+	}
+
+	return cfg, nil
 }
