--- test/aws/aws.go
@@ -1,93 +1,0 @@
-package aws
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/config"
-	"github.com/aws/aws-sdk-go-v2/credentials"
-	"github.com/aws/aws-sdk-go-v2/service/sns"
-	"github.com/aws/aws-sdk-go-v2/service/sqs"
-)
-
-type SQSAPI interface {
-	CreateQueue(ctx context.Context, params *sqs.CreateQueueInput, optFns ...func(*sqs.Options)) (*sqs.CreateQueueOutput, error)
-	ReceiveMessage(ctx context.Context, params *sqs.ReceiveMessageInput, optFns ...func(*sqs.Options)) (*sqs.ReceiveMessageOutput, error)
-}
-
-type SNSAPI interface {
-	CreateTopic(ctx context.Context, params *sns.CreateTopicInput, optFns ...func(*sns.Options)) (*sns.CreateTopicOutput, error)
-}
-
-// CreateSNSAPI helper function.
-func CreateSNSAPI(region, endpoint string) (*sns.Client, error) {
-	cfg, err := createConfig(sns.ServiceID, region, endpoint)
-	if err != nil {
-		return nil, err
-	}
-
-	api := sns.NewFromConfig(cfg)
-
-	return api, nil
-}
-
-// CreateSNSTopic helper function.
-func CreateSNSTopic(api SNSAPI, topic string) (string, error) {
-	out, err := api.CreateTopic(context.Background(), &sns.CreateTopicInput{
-		Name: aws.String(topic),
-	})
-	if err != nil {
-		return "", fmt.Errorf("failed to create topic %s: %w", topic, err)
-	}
-
-	return *out.TopicArn, nil
-}
-
-// CreateSQSAPI helper function.
-func CreateSQSAPI(region, endpoint string) (*sqs.Client, error) {
-	cfg, err := createConfig(sqs.ServiceID, region, endpoint)
-	if err != nil {
-		return nil, err
-	}
-
-	api := sqs.NewFromConfig(cfg)
-
-	return api, nil
-}
-
-// CreateSQSQueue helper function.
-func CreateSQSQueue(api SQSAPI, queueName string) (string, error) {
-	out, err := api.CreateQueue(context.Background(), &sqs.CreateQueueInput{
-		QueueName: aws.String(queueName),
-	})
-	if err != nil {
-		return "", err
-	}
-
-	return *out.QueueUrl, nil
-}
-
-func createConfig(awsServiceID, awsRegion, awsEndpoint string) (aws.Config, error) {
-	customResolver := aws.EndpointResolverWithOptionsFunc(func(service, region string, options ...interface{}) (aws.Endpoint, error) {
-		if service == awsServiceID && region == awsRegion {
-			return aws.Endpoint{
-				URL:           awsEndpoint,
-				SigningRegion: awsRegion,
-			}, nil
-		}
-		// returning EndpointNotFoundError will allow the service to fallback to it's default resolution
-		return aws.Endpoint{}, &aws.EndpointNotFoundError{}
-	})
-
-	cfg, err := config.LoadDefaultConfig(context.TODO(),
-		config.WithRegion(awsRegion),
-		config.WithEndpointResolverWithOptions(customResolver),
-		config.WithCredentialsProvider(aws.NewCredentialsCache(credentials.NewStaticCredentialsProvider("test", "test", ""))),
-	)
-	if err != nil {
-		return aws.Config{}, fmt.Errorf("failed to create AWS config: %w", err)
-	}
-
-	return cfg, nil
-}
